Qt ui code: fix qreal type problems on ARM platform
fixing problems in qt ui code comming from the fact that qreal type is float on
ARM platform and double on other platforms, which causes build errors on ARM
(not all casts are probably strictly necessary)

Bug: 12483
Reviewed-on: https://code.wireshark.org/review/15723

+++ ui/qt/lbm_uimflow_dialog.cpp
@@ -364,14 +364,14 @@ void LBMUIMFlowDialog::vScrollBarChanged(int value)
 
 void LBMUIMFlowDialog::xAxisChanged(QCPRange range)
 {
-    m_ui->horizontalScrollBar->setValue(qRound(range.center() * 100.0));
-    m_ui->horizontalScrollBar->setPageStep(qRound(range.size() * 100.0));
+    m_ui->horizontalScrollBar->setValue(qRound(qreal(range.center() * 100.0)));
+    m_ui->horizontalScrollBar->setPageStep(qRound(qreal(range.size() * 100.0)));
 }
 
 void LBMUIMFlowDialog::yAxisChanged(QCPRange range)
 {
-    m_ui->verticalScrollBar->setValue(qRound(range.center() * 100.0));
-    m_ui->verticalScrollBar->setPageStep(qRound(range.size() * 100.0));
+    m_ui->verticalScrollBar->setValue(qRound(qreal(range.center() * 100.0)));
+    m_ui->verticalScrollBar->setPageStep(qRound(qreal(range.size() * 100.0)));
 }
 
 void LBMUIMFlowDialog::diagramClicked(QMouseEvent * event)
